Corrected details for the 'configFile' param
Corrected all details related to the 'configFile' parameter. The current documents refer to a .JSON file which does not exist, the script expects a .YAML file (conf_local.yaml' or 'conf_cloud.yaml). 

Corrected the examples to refer to the 'conf_local.yaml' or 'conf_cloud.yaml' files included in this repo.