Implementing TenantId in events: UserCheckoutAcceptedIntegrationEvent, OrderStartedIntegrationEvent, OrderStatusChangedToSubmittedIntegrationEvent, GracePeriodConfirmedIntegrationEvent, OrderStatusChangedToAwaitingValidationIntegrationEvent.
Additionally the TenantId in the events are made use of the EventBusRabbitMQ implementation, where it checks wether the event should be sent to the tenant specific microservice, or handled normally. In addition it is made use of to ensure that events that do not belong to the tenant is not sent to their tenant specific microservice.