Add logic for producing events to kafka eventbus.
For now we have a single topic for all events. Since
we use the eventname as the key for the kafka message,
we have the property that they all get assigned to the same
parition inside kafka and therefore are in-order.

Alternatively one could have multiple topics.

The code for the kafka connection is based on:

https://github.com/confluentinc/confluent-kafka-dotnet/blob/master/examples/Web/KafkaClientHandle.cs